refactor(exporter): modularize data structures and tree exporters into dedicated category files
