+++ features/file_exporter/exporter_trees.c
@@ -1,0 +1,6 @@
+#include "file_exporter.h"
+
+bool export_bst(const char* root_dir, const char* dest_dir)
+{
+    return export_file_pair(root_dir, "bst", "trees", dest_dir, NULL, NULL);
+}
