+++ features/file_exporter/file_exporter.c
@@ -199,28 +199,3 @@ bool export_file_pair(const char* root_dir, const char* base_filename, const cha
 
     return (success_c && success_h);
 }
-
-bool export_sll(const char* root_dir, const char* dest_dir)
-{
-    return export_file_pair(root_dir, "sll", "sll", dest_dir, NULL, NULL);
-}
-
-bool export_dll(const char* root_dir, const char* dest_dir)
-{
-    return export_file_pair(root_dir, "dll", "dll", dest_dir, NULL, NULL);
-}
-
-bool export_bst(const char* root_dir, const char* dest_dir)
-{
-    return export_file_pair(root_dir, "bst", "trees", dest_dir, NULL, NULL);
-}
-
-bool export_circular_queue(const char* root_dir, const char* dest_dir)
-{
-    return export_file_pair(root_dir, "circular_queue", "queue", dest_dir, NULL, NULL);
-}
-
-bool export_stack(const char* root_dir, const char* dest_dir)
-{
-    return export_file_pair(root_dir, "stack", "stack", dest_dir, NULL, NULL);
-}
